+++ CHANGELOG.md
@@ -2,9 +2,11 @@ TODO:
 - Different behavior avr vs pc-based tests to be analyzed
 - #include-file structure needs to be cleaned up
 
-next:
+pre-0.27.0:
 - esp32: StepperDemo make p<n> command work again
 - esp32: check step pin is a valid gpio output pin
+- esp32: Support rmt device with two steppers
+- new return code for `addQueueEntry()`: `AQE_DEVICE_NOT_READY`
 
 0.26.0:
 - avr: Max stepper speed depending on number of assigned steppers: